replace unsafe_convert methods with safe cconvert ones
This is the correct way to define most conversions, since the unsafe
logic should typically be isolated to a few areas, and everywhere else
just defines conversions to it. This ensures the root is preserved even
if the user later makes unexpected changes to the array object (although
the length might have later become inconsistent at that point, so it is
not a guaranteed fix).