Make precompile files relocatable (#49866)
String replacement with `@depot` when serializing out happens with any
paths that are located inside a `DEPOT_PATH` (first match wins). If no
match, then we emit the absolute file path as before. Right now we only
emit one token `@depot`.

String replacement of `@depot` when loading happens now on a `.ji` file
basis and only if all the listed include dependencies can be resolved to
files located in one and the same depot on `DEPOT_PATH` (again, first
match wins). If we can't resolve, then the cache is invalided with
`stale_cachefile`.