atexit: move hook before precompile output
To show how this is used, this updates the profile_printing_listener
background job to use this mechanism.

Also ensure mktemp and mktempdir cleanup even on unexpected exit:
Previously this was relying on all Tasks running to completion, which is
not a good assumption. Add the atexit hook, so that even if they do not
run to completion, they still get cleaned up.