More detail on locks
After reading the documentation on multi-threading, I did not catch the importance of locks and how to use them. This was made clear to me in [this](https://discourse.julialang.org/t/help-find-best-approach-to-streaming-data-on-raspberry-pi/105130/5) discourse thread.

In this PR, I have attempted to add the detail I was missing. It probably needs a second set of eyes to shorten the language and use a style that is more in line with the rest of the documentation.