Significant reordering and rewriting
I ended up significantly rewriting the section on locks, creating a seperate section for locks, reorder things, and add references between sections. I feel like the result is a document with better conceptual structure. 

I also rewrote a section on how to write a thread-safe with `@threads`, as I found the original sentence(s) extremely complicated and hard to read.