[REPL] Fix shell completion error on unreadable symlink (#51851)
In macOS there is a file, `/usr/sbin/weakpass_edit` that is a symlink to
a directory that is only readable by `root`. Our REPL shell completion
attempts to call `isfile()` when entering the shell mode and pressing
`w`, throwing an ugly error into the REPL.

This PR fixes the root cause (skipping files that fail the `isfile()`
condition) and adds a test that synthesizes an analogous condition.