REPL: Try to fix intermittent hang during precompile (#51891)
I was intermittently observing the REPL precompile process not
finishing. What I believe was happening is the following:

1. The last line of the precompile script is
`cd("complete_path\t\t$CTRL_C`
2. As soon as child julia sees the `CTRL_C`, the prompt is terminated,
and a new prompt is echod.
3. The parent julia tries to complete the line by sending `\n`, and
immediately returns, because the child has already written a new prompt.
4. The child reads the `\n` and enters raw mode (ignoring things like
^D).
5. The parent tries to write `^D` to complete the process, but because
the child is still processing the `\n`, this `^D` is ignored and the
process hangs.

Try to fix this by not writing the superfluous `\n` if the precompile
line ends in `^C`.