Fix :noshift construction of an empty SubString
An empty constructed SubString is over the range 0:0. The :noshift
constructor is intended to allow for copying the raw range of a
SubString without modification. However, when trying to copy an empty
SubString, the 0:0 range runs afoul of the boundscheck. This can be
adjusted for by eliding the boundscheck when the range is 0:0.