Fix diagnostics printing when `pwd()` doesn't exist (JuliaLang/JuliaSyntax.jl#373)
Computing the file URL for pretty printing using `abspath` can fail when
the current working directory doesn't exist. This is a quick fix for
this issue (which is incredibly confusing/disruptive if the user does
manage to enter a nonexistant working directory!)

A more complete fix would avoid ever looking at the working directory
when printing diagnostics, instead requiring the caller to pass in a
richer definition of the "location of source code" than the mere file
name as a string. However getting something sensible working there is
(a) breaking and (b) unclear on may details. So just patching this up
quickly seems good for now.