lowering: Fix accidental clearing of never-undef flag in closure conversion pass
This code dates back all the way to the original closure conversion code.
At the time, the 0x4 bit of the vinfo was used for tracking whether a
variable is assigned in an inner function, which apparently needed to
be cleared during closure conversion. However, shortly afterwards,
the bit was removed [3] and subsequently re-used for never-undef [4].
However, the magic number in the closure conversion pass was never
updated, causing the new never-undef flag to be accidentally cleared
unconditionally.

[1] https://github.com/JuliaLang/julia/commit/a1ee8676f00fc14bd67aaf7eb20ce46ff99f322e
[2] https://github.com/JuliaLang/julia/blob/a1ee8676f00fc14bd67aaf7eb20ce46ff99f322e/src/ast.scm#L217-L218
[3] https://github.com/JuliaLang/julia/commit/7e7d74332953497cd0dba278aa243f3efc284cee
[4] https://github.com/JuliaLang/julia/commit/3d8f8d14787e697fab4543e1bca1ce4519977281

Co-authored-by: Jeff Bezanson <jeff@juliahub.com>