squash! indexing ranges with empty ranges
Reviewing the work here, it seems undesirable to me now to actually to
make the majority of this change. We already define too much of what the
value of start should be, and trying to assume that one(T):zero(T) is in
range for T<:Integer is not necessarily ideal either.

This keeps some minor semi-related changes though:
 - ensured this only happens when first(r)+first(s)*st == typemax
 - if the range is supposed to be empty, then use a step of eps=1 (as
   steprange_last_empty will do later), to reduce the size of the set of
   values that overflow and become !isempty or error to 1 value instead of
   step values
 - Handle a convoluted case of StepRangeLen{Bool} indexing into a Vector
   returning the element from the wrong end

```
julia> (typemin(Int)+1:100:zero(Int))[1:0] |> length
184467440737095516 # old
0 # new
```