REVERT ME: revert the re-introduction of PartialQuickSort
There are some pretty compelling reasons to use PartialQuickSort instead of ScratchQuickSort as a base case for BracketedSort.
However, PartialQuickSort can segfault for invalid but seemingly innocuous lt functions.
I don't want to talk about introducing segfaults to the default sorting algorithms or introducing a regression to PartialQuickSort in this PR
Revert this commit it you want `partialsort!(::Vector{Int}, ::Int)` not to allocate.