interpreter: Fix stacktrace for global incorrectly assumed to be in phi block (#51990)
We don't fully set up the interpreter state when evaluating the phi
block, because we expect all statements to be valid in value-position
(using the IR definition of value position). However, the interpreter
was overapproximating the size of the phi-block, leading to the
possibility that the evaluation may throw. Correct that by truncating
the phi block to the actual last phi. Will be tested in existing tests
after #51970.