codegen: remove UB from uninitialized bitstypes in new
In the time since the creation of issue #26764, there _is_ now 'a way to
say to llvm "I don't care what this value is, but it always has to be
the same"', so we can use that to instruct LLVM to not give us undefined
behavior when users are using uninitialized memory. There should not be
an impact if users were already avoiding this paradigm and are fully
initializing their structs.

Fixes #26764