stop passing `arginfo` to `builtin_effects`
`builtin_effects` no longer requires the `arginfo` parameter, as
`getfield_effects` doesn't depend on analyzing `fargs` for
`:boundscheck` information anymore. The equivalent analysis is now
done as the post-opt analysis.

Given that `:boundscheck` expressions are now lowered to an independent
statement, even if we need to analyze `fargs` within `builtin_effects`,
we would likely need to pass `sv::AbsIntState` too in order to allow
it to use `ssa_def_slot`. For now, it's more efficient to just eliminate
the passing of `arginfo` since it is not necessary.