post-opt-analysis: use EA to refine `:effect_free` (#51494)
This commit is aiming to integrate EA into the Base compiler pipeline by
using it during the post-opt analysis to refine `:effect_free`
information. In doing so, this also generalizes `argescapes` field of
`Union{InferenceResult,CodeInstance}` to
`analysis_results::AnalysisResults` so that it can hold results of
multiple post-optimization analyses, where `AnalysisResults` is designed
to be linked-list like data structure. This is because an external
`AbstractInterpreter`, like `EscapeAnalyzer`, might perform several
post-optimization analyses. Honestly speaking, however, I’m not
completely satisfied with this solution yet. It might make more sense to
require a single post-optimization analysis for every
`AbstractInterpreter` always, as like what we do for the other fields
like `rettype`.

Co-authored-by: Keno Fischer <keno@juliacomputing.com>