build: disable DT_TEXTREL warning for i686 (#52211)
This warning was distracting and annoying me, so I looked into whether
it was a bug. Turns out the warning is harmless (loading the object just
will make a copy of the ~2kb text field into memory). So add the code to
disable it, either with different assembly or different linker flags.
Default to continuing to use the same assembly.