Tag CodeInstance for external abstract interpreter
External abstract interpreters need to be able to take part in
invalidation, as well as cross-session inference result caching.

MethodInstance.cache contains a set of CodeInstances which hold the
inference cache results. Instead of haveing external abstract
interpreters create detached CodeInstances we attach them the the
MethodInstance they were created from. This solves the invalidation
problem and the cross-session cache problem in one go, but requires
that we tag/mark the CodeInstance with a token to remember which
abstract interpreter created them.

`nothing` is reserved for the native interpreter and all other
interpreters can use values that compare with `jl_egal`.

To perform a cache-lookup we need the current abstract interpreter
token.

Co-authored-by: Shuhei Kadowaki <40514306+aviatesk@users.noreply.github.com>