effects: refine `:nothrow` when `exct` is known to be `Bottom`
Implements effects refinement that goes in the reverse direction of what
was implemented in 8dd0cf5.

In the discussion at JuliaLang/julia#52268, there's a debate on how to
deal with exceptions like `StackOverflow` or `InterruptException` that
are caused by environment. The original purpose of #52268 was to fix
issues where methods that could cause a StackOverflow were mistakenly
refined to `:nothrow`. But now it feels like it's a bit weird to model
such environment-dependent exceptions as `:nothrow` in the first place.
If we can exclude errors caused by the environment from `:nothrow`
modeling, those methods might be safely considered `:nothrow`. To
prevent the concrete evaluation of methods that may cause StackOverflow,
it's enough to taint its `:terminates` after all.

Anyway, now the compiler is able to prove `:nothrow`-ness of
the following frame:
```julia
function getindex_nothrow(xs::Vector{Int}, i::Int)
    try
        return xs[i]
    catch err
        err isa BoundsError && return nothing
        rethrow(err)
    end
end
```