Fix off-by-2 in codegen
There were two separate off-by-1's in the codegen code that is
trying to detect assignmnents to slots inside try/catch regions.

First, it was asking to include the value of the catch label,
which is actually the first statement *not* in the try region.
Second, there was a confusion of 0 and 1 based indexing in
the iteration bounds. The end result of this was that the code
was also looking at the first two statements of the catch region.

This didn't used to be a problem (other than a potentially
over-conservative marking of some slots as volatile), because
our catch blocks always had at least two statements (a :leave
and a terminator), but with the `:leave` change, it is possible
to have catch blocks with only one statement. If these happened
to be at the end of the function, things would blow up.

As a side node, this code isn't particularly sound, because it
assumes that try/catch regions are lexical, which they are not.
The assumption happens to work out ok for the code we generate
in the frontend and optimized IR doesn't have slots, so we don't
use this code, but it is not in general sound.