Bail out of constprop if we already know it's gonna throw an error (#52293)
While looking at some other things, I noticed that we call concrete
evaluation on every `error("foo")`-like call just to find out again that
it errors. which is obviously wasteful. In the future we may want to
model exception-`:consistent`-cy in which case there could be some
additional `exct` refinement here, but we're not there yet.

---------

Co-authored-by: Kristoffer Carlsson <kcarlsson89@gmail.com>