Automatically :leave the exception frame on the catch edge (#52245)
Right now, we require a :leave expression at both the end of a try
region and as the first expression in the catch block. This doesn't
really make sense. Throwing the exception should leave the exception
frame implicitly. This isn't a huge saving, but does save one IR node
(and generated call in the native code) per try/catch block.